pyproject.toml
[project]
name = "fastapi-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-multipart>=0.0.20",
    "ruff>=0.11.7",
    "sqlalchemy>=2.0.40",
]

[tool.ruff]
line-length = 79  # Максимальная длина строки по PEP8
indent-width = 4  # Использование пробелов для отступов
target-version = "py312"  # Укажи версию Python, если нужно

[tool.ruff.lint]
select = ["E", "F", "W"]  # Pycodestyle (E), Pyflakes (F), предупреждения (W)
ignore = ["E501"]  # Отключаем ограничение длины строки, если используешь Black

[tool.ruff.format]
quote-style = "double"  # PEP8 рекомендует двойные кавычки для строк
indent-style = "space"  # Пробелы вместо табуляции
skip-magic-trailing-comma = false  # Форматирование списков по PEP8

